Address Sam's review comments

+++ contracts/common/LibWormholeGUID.sol
@@ -7,8 +7,8 @@ struct WormholeGUID {
   address receiver;
   address operator;
   uint128 amount;
-  uint64 nonce;
-  uint64 timestamp;
+  uint80 nonce;
+  uint48 timestamp;
 }
 
 library WormholeLib {
